requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-address==0.2.5
django-crispy-forms==1.11.2
django-heroku==0.3.1
gunicorn==20.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
